refactor: place-search FSD 레이어 분리 및 라우트 정리
- features/place-search: 공유 자원만 유지
- widgets/place-search: PlaceSearchContent 신규 생성
- Kakao Map 로직 views/log-place-search로 이동
- /log/place → /log/place-search 라우트 변경
- views/log-place-search로 view 이름 변경
