+++ apps/web/app/log/place-search/page.tsx
@@ -1,0 +1,3 @@
+import { SearchPlacePage } from '@/views/log-place-search';
+
+export default SearchPlacePage;
