--- apps/web/app/log/place/page.tsx
@@ -1,3 +1,0 @@
-import { SearchPlacePage } from '@/views/log/place';
-
-export default SearchPlacePage;
