+++ apps/web/src/features/place-search/index.ts
@@ -6,3 +6,7 @@ export {
   getRecentPlaces,
   removeRecentPlace,
 } from './model/recent-places';
+export { default as HighlightText } from './ui/HighlightText';
+export { default as PlaceSearchInput } from './ui/PlaceSearchInput';
+export { default as RecentPlaceItem } from './ui/RecentPlaceItem';
+export { default as RecentPlaceList } from './ui/RecentPlaceList';
