+++ apps/web/src/widgets/place-search/index.ts
@@ -1,0 +1,2 @@
+export type { SearchState } from './model/types';
+export { default as PlaceSearchContent } from './ui/PlaceSearchContent';
