+++ apps/web/src/widgets/place-search/model/types.ts
@@ -1,0 +1,1 @@
+export type SearchState = 'idle' | 'loading' | 'success' | 'empty' | 'error';
